Working copy is about to be up!

+++ jobuser/models.py
@@ -5,11 +5,11 @@ from django.contrib.auth.models import User;
 class JobUser(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE);
     job = models.ForeignKey(Job, on_delete=models.CASCADE);
-    amount_paid = models.IntegerField(default=0);
+    amount_paid = models.FloatField(default=0.0);
 
 class Pledge(models.Model):
     jobuser = models.ForeignKey(JobUser, on_delete=models.CASCADE);
-    amount = models.IntegerField(editable=False, null=True, blank=True);
+    amount = models.FloatField(editable=False, null=True, blank=True);
     date = models.DateTimeField(auto_now_add=True);
     
 class Pay(models.Model):
